requirements.txt
streamlit
fastapi
requests
newspaper3k
transformers
gtts
rake-nltk
dotenv
streamlit==1.41.1
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
python-dotenv==1.0.1
newspaper3k==0.2.8
beautifulsoup4==4.12.3
transformers==4.48.0
torch==2.6.0+cu118
rake-nltk==1.0.6
nltk==3.9.1
gtts>=2.3.2
python-multipart==0.0.20
pandas==2.2.3
numpy==1.26.4